requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.27.0
python-dotenv>=1.0.0
google-cloud-texttospeech>=2.16.0
llama-index>=0.10.0
llama-index-llms-ollama>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
sentence-transformers>=2.2.2
sounddevice>=0.4.0
numpy>=1.24.0
PyQt6>=6.5.0
